gym_exps/requirements/base.txt
../deeprl
gymnasium[classic-control,mujoco]==0.27.*
hydra-core==1.3.*
hydra-zen==0.8.*
hydra-joblib-launcher==1.2.*
hydra-colorlog==1.2.*
torch-tb-profiler  # PyTorch Profiler TensorBoard Plugin
